word_processing_scripts/requirements.txt
# Core NLP dependencies
spacy>=3.4.0
wordfreq>=3.0.0

# Configuration management
PyYAML>=6.0

# Optional advanced NLP (for enhanced Spanish analysis)  
stanza>=1.4.0
transformers>=4.52.0
torch>=1.10.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
mypy>=1.0.0
